# contracts/whatsapp-gateway/openapi.yaml
openapi: 3.1.0
info:
  title: Enpii WhatsApp Gateway
  version: 1.0.0
  description: Internal API contract. Evolution credentials never cross this boundary.
  license:
    name: Proprietary
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /messages/text:
    post:
      summary: Send text message
      operationId: sendText
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TextMessage'
      responses:
        '202':
          $ref: '#/components/responses/SendResult'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
        '502':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
  /messages/media:
    post:
      summary: Send media message
      operationId: sendMedia
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MediaMessage'
      responses:
        '202':
          $ref: '#/components/responses/SendResult'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '422':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
        '502':
          $ref: '#/components/responses/Error'
        '503':
          $ref: '#/components/responses/Error'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 200
        pattern: '^[A-Za-z0-9._:-]+$'
  schemas:
...
    TextMessage:
      type: object
      additionalProperties: false
      required: [instance_id, to, text]
      properties:
        instance_id:
          type: string
          minLength: 1
          maxLength: 100
        to:
          type: string
          pattern: '^\+?[1-9]\d{7,14}$'
        text:
          type: string
          minLength: 1
          maxLength: 4096
    MediaMessage:
      type: object
      additionalProperties: false
      required: [instance_id, to, media_url]
      properties:
        instance_id:
          type: string
          minLength: 1
          maxLength: 100
        to:
          type: string
          pattern: '^\+?[1-9]\d{7,14}$'
        media_url:
          type: string
          format: uri
          pattern: '^https://.+'
        caption:
          type: [string, 'null']
          maxLength: 4096
        filename:
          type: [string, 'null']
          maxLength: 255
    SendResult:
      type: object
      additionalProperties: false
      required: [message_id, status]
      properties:
        message_id:
          type: string
        status:
          type: string
          enum: [accepted, sent]
    Error:
      type: object
      additionalProperties: false
      required: [code, message, request_id, retryable]
      properties:
        code:
          type: string
          pattern: '^[A-Z0-9_]+$'
        message:
          type: string
          maxLength: 300
        request_id:
          type: string
        retryable:
          type: boolean
  responses:
...
    SendResult:
      description: Accepted message
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SendResult'
    Error:
      description: Safe error envelope; never contains provider response or credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'